requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
PyPDF2==3.0.1
python-docx==1.1.0
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
faiss-cpu==1.7.4
numpy==1.24.3
gunicorn==21.2.0